refactor: MapStruct 오류 해결
