+++ src/test/java/team03/monew/service/comments/CommentServiceTest.java
@@ -31,11 +31,11 @@ import team03.monew.module.article.dto.CommentCreateRequest;
 import team03.monew.module.article.dto.CommentDto;
 import team03.monew.module.article.dto.CommentLikeDto;
 import team03.monew.module.article.dto.CommentUpdateRequest;
+import team03.monew.module.comments.entity.QComment;
 import team03.monew.module.common.dto.CursorPageResponse;
 import team03.monew.module.comments.entity.Comment;
 import team03.monew.module.comments.entity.CommentLike;
 import team03.monew.module.article.entity.Article;
-import team03.monew.entity.comments.QComment;
 import team03.monew.module.user.entity.User;
 import team03.monew.module.user.entity.User.Role;
 import team03.monew.module.comments.mapper.CommentMapper;
